NLS: Use language specific response characters
The characters used to respond to a question can change according
what language the program is being run with, for instance the prompt
"Y/N" in English, would be "S/N" in Spanish and "J/N" in German. Use
the new kitten function to parse these characters from the prompt msg
so that the proper keys can be used to respond. [Fixes #11]